revert reorganization of project files
