+++ internal/pkg/scaffold/build_dockerfile.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package project
+package scaffold
 
 import (
 	"path/filepath"
