+++ internal/pkg/scaffold/gopkgtoml_test.go
@@ -12,30 +12,28 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package project
+package scaffold
 
 import (
-	"fmt"
+	"testing"
 
-	"github.com/operator-framework/operator-sdk/internal/pkg/scaffold/input"
-	"github.com/operator-framework/operator-sdk/internal/pkg/scaffold/internal/deps"
+	"github.com/operator-framework/operator-sdk/internal/util/diffutil"
 )
 
-const GopkgTomlFile = "Gopkg.toml"
-
-type GopkgToml struct {
-	input.Input
-}
-
-func (s *GopkgToml) GetInput() (input.Input, error) {
-	if s.Path == "" {
-		s.Path = GopkgTomlFile
+func TestGopkgtoml(t *testing.T) {
+	s, buf := setupScaffoldAndWriter()
+	err := s.Execute(appConfig, &GopkgToml{})
+	if err != nil {
+		t.Fatalf("Failed to execute the scaffold: (%v)", err)
 	}
-	s.TemplateBody = gopkgTomlTmpl
-	return s.Input, nil
+
+	if gopkgtomlExp != buf.String() {
+		diffs := diffutil.Diff(gopkgtomlExp, buf.String())
+		t.Fatalf("Expected vs actual differs.\n%v", diffs)
+	}
 }
 
-const gopkgTomlTmpl = `# Force dep to vendor the code generators, which aren't imported just used at dev time.
+const gopkgtomlExp = `# Force dep to vendor the code generators, which aren't imported just used at dev time.
 required = [
   "k8s.io/code-generator/cmd/deepcopy-gen",
   "k8s.io/code-generator/cmd/conversion-gen",
@@ -110,11 +108,3 @@ required = [
     name = "k8s.io/gengo"
     non-go = false
 `
-
-func PrintDepGopkgTOML(asFile bool) error {
-	if asFile {
-		_, err := fmt.Println(gopkgTomlTmpl)
-		return err
-	}
-	return deps.PrintDepGopkgTOML(gopkgTomlTmpl)
-}
